requirements_gradio.txt
# Gradio Web界面依赖

gradio>=3.40.0
pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0
google-genai>=0.1.0
